+++ src/fundamental/tpm2-pcr.h
